+++ packages/xchain-bitcoin/src/const.ts
@@ -1,6 +1,7 @@
 import { ExplorerProvider, Network, UtxoOnlineDataProviders } from '@xchainjs/xchain-client'
 import { Asset } from '@xchainjs/xchain-util'
 import {
+  BitgoProvider,
   BlockcypherNetwork,
   BlockcypherProvider,
   HaskoinNetwork,
@@ -19,7 +20,7 @@ export const MIN_TX_FEE = 1000
 export const BTC_DECIMAL = 8
 
 export const LOWER_FEE_BOUND = 1
-export const UPPER_FEE_BOUND = 500
+export const UPPER_FEE_BOUND = 1_000
 export const BTC_SYMBOL = '₿'
 export const BTC_SATOSHI_SYMBOL = '⚡'
 
@@ -114,8 +115,23 @@ const mainnetBlockcypherProvider = new BlockcypherProvider(
   BlockcypherNetwork.BTC,
   process.env.BLOCKCYPHER_API_KEY || '',
 )
+
 export const BlockcypherDataProviders: UtxoOnlineDataProviders = {
   [Network.Testnet]: testnetBlockcypherProvider,
   [Network.Stagenet]: mainnetBlockcypherProvider,
   [Network.Mainnet]: mainnetBlockcypherProvider,
+}
+
+//======================
+// Bitgo
+//======================
+const mainnetBitgoProvider = new BitgoProvider({
+  baseUrl: 'https://app.bitgo.com',
+  chain: BTCChain,
+})
+
+export const BitgoProviders: UtxoOnlineDataProviders = {
+  [Network.Testnet]: undefined,
+  [Network.Stagenet]: mainnetBitgoProvider,
+  [Network.Mainnet]: mainnetBitgoProvider,
 }