+++ packages/xchain-utxo-providers/src/providers/blockcypher/blockcypher-data-provider.ts
@@ -1,6 +1,8 @@
 import PromisePool from '@supercharge/promise-pool'
 import {
   Balance,
+  FeeOption,
+  FeeRates,
   Tx,
   TxHash,
   TxHistoryParams,
@@ -116,6 +118,24 @@ export class BlockcypherProvider implements UtxoOnlineDataProvider {
       throw error
     }
   }
+
+  /**
+   * Returns a fee rate estimation from Blockcypher API service.
+   * @returns {FeeRates} Estimated fee rates
+   */
+  async getFeeRates(): Promise<FeeRates> {
+    const chainResponse = await blockcypher.getBlockchainData({
+      baseUrl: `${this.baseUrl}/${this.blockcypherNetwork}`,
+      apiKey: this._apiKey,
+    })
+
+    return {
+      [FeeOption.Average]: chainResponse.low_fee_per_kb / 1000,
+      [FeeOption.Fast]: chainResponse.medium_fee_per_kb / 1000,
+      [FeeOption.Fastest]: chainResponse.high_fee_per_kb / 1000,
+    }
+  }
+
   private mapTransactionToTx(rawTx: Transaction): Tx {
     return {
       asset: this.asset,