+++ packages/xchain-utxo-providers/src/providers/sochainv3/sochain-data-provider.ts
@@ -1,5 +1,6 @@
 import {
   Balance,
+  FeeRates,
   Tx,
   TxHash,
   TxHistoryParams,
@@ -183,6 +184,13 @@ export class SochainProvider implements UtxoOnlineDataProvider {
       console.error(error)
       throw error
     }
+  }
+
+  /**
+   *  @throws {Error} Method not implemented.
+   */
+  getFeeRates(): Promise<FeeRates> {
+    throw new Error('Method not implemented.')
   }
 
   private mapUTXOs(utxos: AddressUTXO[]): UTXO[] {