+++ packages/xchain-utxo-providers/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.2.8 (2023-11-21)
+
+## Update
+
+- Bitgo provider
+- Get fee rates endpoint for providers
+
 # v0.2.7 (2023-11-16)
 
 ## Update